Deka

+++ app/Http/Controllers/Main/SiteController.php
@@ -48,10 +48,14 @@ class SiteController extends Controller
             })
             ->addColumn('action', function ($data) {
                 $btn = '<div class="d-flex">';
-                $btn .= '<a href="javascript:void(0);" class="btn-edit-data btn-sm me-1 shadow me-2" title="Edit Data" onclick="editData(\'' . $data->id . '\')"><i class="ti ti-pencil"></i></a>';
-                $btn .= '<a href="javascript:void(0);" class="btn-delete-data btn-sm shadow" title="Hapus Data" onclick="deleteData(\'' . $data->id . '\')"><i class="ti ti-trash"></i></a>';
+                if (!auth()->user()->hasRole('Read only')) {
+                        $btn .= '<a href="javascript:void(0);" class="btn-edit-data btn-sm me-1 shadow me-2" title="Edit Data" onclick="editData(\'' . $data->id . '\')"><i class="ti ti-pencil"></i></a>';
+                    }
+                if (!auth()->user()->hasRole('Read only')) {
+                        $btn .= '<a href="javascript:void(0);" class="btn-delete-data btn-sm shadow" title="Hapus Data" onclick="deleteData(\'' . $data->id . '\')"><i class="ti ti-trash"></i></a>';
+                    }
                 $btn .= '</div>';
-            
+
                 return $btn;
             })
             ->escapeColumns([])
@@ -98,7 +102,7 @@ class SiteController extends Controller
         try {
             return $this->atomic(function () use ($data) {
                 $data = Site::create($data);
-                
+
                 return response()->json([
                     'status' => true,
                     'message' => 'Data berhasil ditambahkan!',
